Created a hilcode directory storing hil components (probably to split up even further later).
